Add files via upload
You are employed as a data scientist at SussexBudgetProductions. The last movie produced by the company was a comedy-action-thriller with a budget of £500K. It failed at the box office and only grossed £100K. So to keep the production company in business, the next film needs to make money. The CEO thinks they can persuade investors to give £1.5 million in funding over the next year, but needs to ensure a good return. Therefore, they have asked you to analyse the IMDB data provided in the movie metadata.csv Download metadata.csvfile and come up with an actionable plan. You should assume that all budgets in the data file have been converted to GBP.